requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
fsspec==2024.6.1
googleapis-common-protos==1.63.2
grpcio==1.65.4
gunicorn==23.0.0
h11==0.14.0
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
lz4==4.3.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
packaging==24.1
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
protobuf==4.25.4
protoc-gen-openapiv2==0.0.1
psycopg2==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
setuptools==72.2.0
sqlparse==0.5.1
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
